add guard and spork for live test updates
To use this, run the command "guard" in a separate terminal window. Specs will be run whenever a file is changed, and apparently will keep running as long as there are failing tests.

Spork just makes them run more efficiently. But remember to restart spork (or in this case guard) when you change files in the prefork loading (such as routes.rb) or they might not be recognizedi, and that could give you false positive or negative results.